rename context to Step context
